+++ src/main/java/com/example/cafecontrolsystem/entity/Payment.java
@@ -11,7 +11,7 @@ public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payment_id")
-    private Integer paymentId;
+    private Long paymentId;
     
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "sale_id")
