+++ src/main/java/com/example/cafecontrolsystem/entity/SaleDetail.java
@@ -11,7 +11,7 @@ public class SaleDetail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "detail_id")
-    private Integer detailId;
+    private Long detailId;
     
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "sale_id")
